requirements.txt
fastapi==0.95.1
mangum==0.17.0
uvicorn==0.21.1
fastapi-mail==1.2.7
httpx==0.24.0
pytest==7.3.1
python-dotenv==1.0.0